chore(agent-tunnel): fix CI, relocate tests, elaborate override rationale
CI fix:
- enrollment.rs: replace `time::OffsetDateTime::now_utc()` with
  `std::time::SystemTime::now()` — the `time` crate is Windows-only in
  devolutions-agent's Cargo.toml, so the previous code broke the Linux
  lint and test jobs. No dependency added; one fewer path too.

Test relocation:
- `crates/agent-tunnel/src/integration_test.rs` (QUIC E2E +
  domain-routing E2E) moved to
  `devolutions-gateway/tests/agent_tunnel_integration.rs`. Integration
  tests belong in the gateway's dedicated `tests/` folder, not inside
  a library crate's `src/`. Imports swapped from `super::cert` /
  `super::listener` to `agent_tunnel::cert` / `agent_tunnel::`.
- `devolutions-agent/src/main.rs` inline `mod tests { ... }` block
  extracted to `devolutions-agent/src/cli_tests.rs`. Kept as a child
  module (`#[cfg(test)] mod cli_tests;`) because the tests use the
  private `UpCommand` / `parse_up_command_args` — the alternative
  (move to `tests/` folder) would require exposing binary internals
  or a lib+bin split.

Dev-dep cleanup on agent-tunnel:
- Drop `rustls-pemfile` and `tempfile` from `[dev-dependencies]` —
  both were only used by the integration test that just moved out.
- Trim `tokio` dev-dep to `["macros"]` (drop `"net"` — no TcpListener
  usage in the remaining src tests).
- `rustls` stays in `[dependencies]` — it is unavoidable: Quinn's
  QUIC server TLS config is built on `rustls::ServerConfig`. PR1's
  cleanup dropped `rustls-pemfile` and `x509-parser` (replaced with
  picky); `rustls` itself is the TLS stack, not something we pull in
  for PEM parsing.

`jet_quic_endpoint` rationale:
- Expand the terse "often a container ID in Docker" comment into a
  full explanation on `EnrollmentJwtClaims::jet_quic_endpoint` of why
  the override exists at all: a running process has no way to
  self-discover its externally-reachable address, so the enroll API's
  self-reported `conf.hostname:port` is routinely wrong in
  Docker/K8s, NAT, split-horizon DNS, and HA-behind-LB deployments.
  Only the operator who designed the network knows the right value,
  so it is encoded into the JWT at mint time. The CLI flag takes
  precedence for last-minute corrections without re-issuing a JWT.
- `enroll_agent`'s docstring and the inline comment in the override
  branch now defer to the claim doc instead of re-stating the Docker
  example.

Verification:
- `cargo check --workspace --all-targets` ✅
- `cargo clippy -p agent-tunnel -p agent-tunnel-proto -p devolutions-agent -p devolutions-gateway --all-targets` ✅ 0 warn
- Tests: 55 lib (agent-tunnel 20 / proto 3 / agent 27 / agent bin 5)
  + 25 gateway integration (e2e 2 / registry 13 / routing 10) = 80
  total, all green.
- `cargo +nightly fmt --check` ✅
