+++ crates/agent-tunnel/Cargo.toml
@@ -43,6 +43,4 @@ socket2 = "0.5"
 
 [dev-dependencies]
 base64 = "0.22"
-tempfile = "3"
-rustls-pemfile = "2.2"
-tokio = { version = "1.52", features = ["macros", "net"] }
+tokio = { version = "1.52", features = ["macros"] }
